refactor(questura): decompose listicle collection
